Add simple backend example
