+++ todos_flutter/build.yaml
@@ -1,0 +1,7 @@
+targets:
+  $default:
+    builders:
+      drift_dev:
+        options:
+          sql:
+            dialects: [sqlite, postgres]
