+++ todos_flutter/pubspec.yaml
@@ -32,6 +32,13 @@ dependencies:
   # https://github.com/simolus3/sqlite3.dart/releases?q=sqlite3_flutter_libs&expanded=true
   sqlite3_flutter_libs: 0.5.19+1
 
+  # THese are only used for the backend example, they are not needed to use Eletric
+  # in Flutter apps.
+  postgres: ^3.0.8
+  drift_postgres: ^1.1.0
+  shelf: ^1.4.1
+  shelf_router: ^1.1.4
+
 dev_dependencies:
   build_runner: ^2.4.4
   drift_dev: 2.15.0
